Extract hasTruthyAttribute to shared util
