+++ packages/eslint-plugin/utils/index.js
@@ -9,6 +9,7 @@ const {
 const { getTranslateFunctionArgs } = require( './get-translate-function-args' );
 const { getTextContentFromNode } = require( './get-text-content-from-node' );
 const { getTranslateFunctionName } = require( './get-translate-function-name' );
+const { hasTruthyJsxAttribute } = require( './has-truthy-jsx-attribute' );
 const isPackageInstalled = require( './is-package-installed' );
 
 module.exports = {
@@ -18,5 +19,6 @@ module.exports = {
 	getTranslateFunctionArgs,
 	getTextContentFromNode,
 	getTranslateFunctionName,
+	hasTruthyJsxAttribute,
 	isPackageInstalled,
 };
